refactor: add MockDBSync class corresponding header for testing
