+++ src/modules/sca/tests/mocks/mockdbsync.hpp
@@ -1,0 +1,17 @@
+#pragma once
+
+#include <dbsync.hpp>
+
+#include <string>
+
+std::string MockDBSyncDbFilePath;
+
+class MockDBSync
+{
+public:
+    explicit MockDBSync(
+        const HostType, const DbEngineType, const std::string& dbFilePath, const std::string&, const DbManagement)
+    {
+        MockDBSyncDbFilePath = dbFilePath;
+    }
+};
