+++ src/modules/sca/tests/sca_test.cpp
@@ -5,20 +5,10 @@
 #include <configuration_parser.hpp>
 #include <dbsync.hpp>
 
+#include "mocks/mockdbsync.hpp"
+
 #include <memory>
 #include <string>
-
-std::string MockDBSyncDbFilePath;
-
-class MockDBSync
-{
-public:
-    explicit MockDBSync(
-        const HostType, const DbEngineType, const std::string& dbFilePath, const std::string&, const DbManagement)
-    {
-        MockDBSyncDbFilePath = dbFilePath;
-    }
-};
 
 class ScaTest : public ::testing::Test
 {
